requirements.txt
Scrapy~=2.11.2
itemadapter~=0.9.0
pymongo~=4.10.1
confluent-kafka~=2.10.0
python-dotenv~=1.1.0
redis~=5.0.0
parsel~=1.10.0
openai==0.28.0
elasticsearch==8.11.1
Booktype~=1.5
flask~=3.1.1
flask-cors~=6.0.0